chck resources
