+++ pyftpdlib/test/test_functional.py
@@ -75,16 +75,6 @@ if POSIX:
         pass
 
 
-def check_resources():
-    import psutil
-    import threading
-    p = psutil.Process()
-    ts = threading.enumerate()
-    assert len(ts) == 1, ts
-    children = p.children()
-    assert not children, children
-
-
 class TestFtpAuthentication(unittest.TestCase):
 
     "test: USER, PASS, REIN."
@@ -92,7 +82,6 @@ class TestFtpAuthentication(unittest.TestCase):
     client_class = ftplib.FTP
 
     def setUp(self):
-        check_resources()
         self.server = self.server_class()
         self.server.start()
         self.client = self.client_class(timeout=TIMEOUT)
@@ -108,7 +97,6 @@ class TestFtpAuthentication(unittest.TestCase):
         if not self.dummyfile.closed:
             self.dummyfile.close()
         os.remove(TESTFN)
-        check_resources()
 
     def assert_auth_failed(self, user, passwd):
         self.assertRaisesRegex(ftplib.error_perm, '530 Authentication failed',
@@ -267,7 +255,6 @@ class TestFtpDummyCmds(unittest.TestCase):
     client_class = ftplib.FTP
 
     def setUp(self):
-        check_resources()
         self.server = self.server_class()
         self.server.start()
         self.client = self.client_class(timeout=TIMEOUT)
@@ -277,7 +264,6 @@ class TestFtpDummyCmds(unittest.TestCase):
     def tearDown(self):
         self.client.close()
         self.server.stop()
-        check_resources()
 
     def test_type(self):
         self.client.sendcmd('type a')
@@ -392,7 +378,6 @@ class TestFtpCmdsSemantic(unittest.TestCase):
          'stru', 'type', 'user', 'xmkd', 'xrmd', 'site chmod']
 
     def setUp(self):
-        check_resources()
         self.server = self.server_class()
         self.server.start()
         self.client = self.client_class(timeout=TIMEOUT)
@@ -402,7 +387,6 @@ class TestFtpCmdsSemantic(unittest.TestCase):
     def tearDown(self):
         self.client.close()
         self.server.stop()
-        check_resources()
 
     def test_arg_cmds(self):
         # Test commands requiring an argument.
@@ -457,7 +441,6 @@ class TestFtpFsOperations(unittest.TestCase):
     client_class = ftplib.FTP
 
     def setUp(self):
-        check_resources()
         self.server = self.server_class()
         self.server.start()
         self.client = self.client_class(timeout=TIMEOUT)
@@ -472,7 +455,6 @@ class TestFtpFsOperations(unittest.TestCase):
         safe_remove(self.tempfile)
         if os.path.exists(self.tempdir):
             shutil.rmtree(self.tempdir)
-        check_resources()
 
     def test_cwd(self):
         self.client.cwd(self.tempdir)
@@ -667,7 +649,6 @@ class TestFtpStoreData(unittest.TestCase):
     use_sendfile = None
 
     def setUp(self):
-        check_resources()
         self.server = self.server_class()
         if self.use_sendfile is not None:
             self.server.handler.use_sendfile = self.use_sendfile
@@ -687,7 +668,6 @@ class TestFtpStoreData(unittest.TestCase):
         if self.use_sendfile is not None:
             from pyftpdlib.handlers import _import_sendfile
             self.server.handler.use_sendfile = _import_sendfile() is not None
-        check_resources()
 
     def test_stor(self):
         try:
@@ -989,7 +969,6 @@ class TestFtpRetrieveData(unittest.TestCase):
     use_sendfile = None
 
     def setUp(self):
-        check_resources()
         self.server = self.server_class()
         if self.use_sendfile is not None:
             self.server.handler.use_sendfile = self.use_sendfile
@@ -1011,7 +990,6 @@ class TestFtpRetrieveData(unittest.TestCase):
         if self.use_sendfile is not None:
             from pyftpdlib.handlers import _import_sendfile
             self.server.handler.use_sendfile = _import_sendfile() is not None
-        check_resources()
 
     def test_retr(self):
         data = b'abcde12345' * 100000
@@ -1112,7 +1090,6 @@ class TestFtpListingCmds(unittest.TestCase):
     client_class = ftplib.FTP
 
     def setUp(self):
-        check_resources()
         self.server = self.server_class()
         self.server.start()
         self.client = self.client_class(timeout=TIMEOUT)
@@ -1124,7 +1101,6 @@ class TestFtpListingCmds(unittest.TestCase):
         self.client.close()
         self.server.stop()
         os.remove(TESTFN)
-        check_resources()
 
     def _test_listing_cmds(self, cmd):
         """Tests common to LIST NLST and MLSD commands."""
@@ -1276,7 +1252,6 @@ class TestFtpAbort(unittest.TestCase):
     client_class = ftplib.FTP
 
     def setUp(self):
-        check_resources()
         self.server = self.server_class()
         self.server.start()
         self.client = self.client_class(timeout=TIMEOUT)
@@ -1286,7 +1261,6 @@ class TestFtpAbort(unittest.TestCase):
     def tearDown(self):
         self.client.close()
         self.server.stop()
-        check_resources()
 
     def test_abor_no_data(self):
         # Case 1: ABOR while no data channel is opened: respond with 225.
@@ -1370,7 +1344,6 @@ class TestThrottleBandwidth(unittest.TestCase):
     client_class = ftplib.FTP
 
     def setUp(self):
-        check_resources()
 
         class CustomDTPHandler(ThrottledDTPHandler):
             # overridden so that the "awake" callback is executed
@@ -1402,7 +1375,6 @@ class TestThrottleBandwidth(unittest.TestCase):
             self.dummyfile.close()
         if os.path.exists(TESTFN):
             os.remove(TESTFN)
-        check_resources()
 
     def test_throttle_send(self):
         # This test doesn't test the actual speed accuracy, just
@@ -1466,7 +1438,6 @@ class TestTimeouts(unittest.TestCase):
             self.server.handler.passive_dtp.timeout = 30
             self.server.handler.active_dtp.timeout = 30
             self.server.stop()
-        check_resources()
 
     # Note: moved later.
 
@@ -1628,7 +1599,6 @@ class TestConfigurableOptions(unittest.TestCase):
             self.server.handler.tcp_no_delay = hasattr(socket, 'TCP_NODELAY')
             self.server.stop()
         os.remove(TESTFN)
-        check_resources()
 
     @disable_log_warning
     def test_max_connections(self):
@@ -1809,7 +1779,6 @@ class TestCallbacks(unittest.TestCase):
     TESTFN_2 = TESTFN + "-2"
 
     def setUp(self):
-        check_resources()
 
         class Handler(FTPHandler):
 
@@ -1859,7 +1828,6 @@ class TestCallbacks(unittest.TestCase):
         self.server.stop()
         safe_remove(TESTFN)
         safe_remove(self.TESTFN_2)
-        check_resources()
 
     def read_file(self, text):
         stop_at = time.time() + 1
@@ -2176,7 +2144,6 @@ class TestIPv6MixedEnvironment(unittest.TestCase):
     HOST = "::"
 
     def setUp(self):
-        check_resources()
         self.server = self.server_class((self.HOST, 0))
         self.server.start()
         self.client = None
@@ -2185,7 +2152,6 @@ class TestIPv6MixedEnvironment(unittest.TestCase):
         if self.client is not None:
             self.client.close()
         self.server.stop()
-        check_resources()
 
     def test_port_v4(self):
         def noop(x):
@@ -2253,7 +2219,6 @@ class TestCornerCases(unittest.TestCase):
     client_class = ftplib.FTP
 
     def setUp(self):
-        check_resources()
         self.server = self.server_class()
         self.server.start()
         self.client = self.client_class(timeout=TIMEOUT)
@@ -2263,7 +2228,6 @@ class TestCornerCases(unittest.TestCase):
     def tearDown(self):
         self.client.close()
         self.server.stop()
-        check_resources()
 
     def test_port_race_condition(self):
         # Refers to bug #120, first sends PORT, then disconnects the
@@ -2555,7 +2519,6 @@ class ThreadedFTPTests(unittest.TestCase):
     client_class = ftplib.FTP
 
     def setUp(self):
-        check_resources()
         self.server = self.server_class()
         self.server.start()
         self.client = self.client_class(timeout=TIMEOUT)
@@ -2572,7 +2535,6 @@ class ThreadedFTPTests(unittest.TestCase):
         self.dummy_recvfile.close()
         self.dummy_sendfile.close()
         safe_remove(TESTFN)
-        check_resources()
 
     def test_unforeseen_mdtm_event(self):
         # Emulate a case where the file last modification time is prior
